==> scripts/20_construction_elements_non_geo.Rmd <==
---
title: "Démonstrateur ASPE : Construction des éléments non géographiques"
author: "Pascal Irz"
date: "`r format(Sys.time(), 'Le %d %B %Y')`"
output:
  html_document:
    code_folding: show
    toc: true
    toc_float: true
---
  
```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Chargement des packages et des données

```{r}
library(aspe)
library(tidyverse)
```


```{r, echo = FALSE}
# chargement données déjà prétraitées
load(file = "../processed_data/captures.RData") # chargement données déjà prétraitées

# chargement mesures individuelles
rdata_mei <- misc_nom_dernier_fichier(repertoire = "../../../raw_data",
                                      pattern = "^mei")

load(rdata_mei)

# chargement autres tables
rdata_tables <- misc_nom_dernier_fichier(repertoire = "../../../raw_data",
                                         pattern = "^tables")

load(rdata_tables)
```

Ici, la dernière date de saisie / modification / opération dans la base est le `r format(misc_derniere_date(), '%d %B %Y')`.

# df1 : espèce - opération

### Densité

```{r}
df1 <- captures %>% 
  mutate(dens_ind_1000m2 = 1000 * effectif / ope_surface_calculee) %>% 
  select(-ope_surface_calculee)
```

### Longueur des individus

Calcul des longueurs mini, maxi, moyenne et médiane pour chaque espèce à chaque opération. On ne tient pas compte du type de longueur (fourche ou totale).

```{r}
df1 <- df1 %>% 
  mef_ajouter_stats_taille()
```


### Apparitions / dispartitions d'espèces

Il s'agit ici de caractériser la dynamique - à l'échelle du point de prélèvement - des apparitions et disparitions d'espèces au fil du temps. On appelle ici colonisation le fait qu'une espèce soit capturée lors d'une opération alors qu'elle était absente lors de la précédente. A l'inverse, on désigne par le terme extinction l'absence d'une espèce dans l'échantillonnage alors qu'elle avait été capturée lors de la précédente prospection. 

>NB cette opération sur l'ensemble de la base prend environ 15 minutes.

```{r, echo = FALSE}
load(file = "../processed_data/colo_ext1.RData")
```


```{r, eval = FALSE}
colo_ext1 <- df1 %>%
  rename(lop_effectif = effectif) %>% # nom de la colonne attendu par mef_colo_ext_pops()
  mef_colo_ext_pops()
```

Comme c'est long, il peut être utile de sauvegarder le résultat de ce traitement. 

```{r, eval = FALSE}
save(colo_ext1, file = "../processed_data/colo_ext1.RData")
```

```{r}
df1 <- colo_ext1 %>%
  select(-effectif) %>% 
  left_join(y = df1)
```


### Visualisation d'un échantillon

```{r}
df1 %>%
  sample_n(10) %>%
  DT::datatable() %>%
  DT::formatRound(columns = c('dens_ind_1000m2', 'taille_moy', 'diam_point'),
                  digits = 2)
```


# df2 : peuplement - opération

### Agrégation

```{r}
df2 <- df1 %>% 
  group_by(pop_id,
           ope_id,
           annee) %>% 
    summarise(richesse = n_distinct(esp_code_alternatif),
              effectif = sum(effectif, na.rm = TRUE),
              dens_ind_1000m2 = sum(dens_ind_1000m2,
                                    na.rm = TRUE)) %>% 
  ungroup()
```

### Apparitions / dispartitions d'espèces

Dénombrement des colonisations / extinction par point chaque année.

```{r}
colo_ext2 <- colo_ext1 %>%
  group_by(annee, pop_id) %>%
    summarise(n_colo = sum(col_ext == "colonisation"),
              n_ext = sum(col_ext == "extinction")) %>% 
  ungroup()
```

Ajout au dataframe.

```{r}
df2 <- df2 %>% 
  left_join(colo_ext2)
```

### IPR

Rajout des valeurs IPR, des métriques et des classes de qualité. Les autres variables (valeurs théoriques et observées des métriques) sont supprimées.

```{r}
df2 <- df2 %>%
  mef_ajouter_metriques() %>% 
  mef_ajouter_ipr() %>% 
  select(-(ner_theorique:dti_observe))
```

# df3 : espece - n opérations

### Complétion du df

Comme les données enregistrées dans la base sont les captures, les effectifs ne sont jamais nuls. Il faut donc compléter le tableau de données par espèce pour distinguer les années où il y a eu prospection sans capture de l'espèce des années où il n'y a pas eu prospection.

```{r}
df3 <- df1 %>% 
  group_by(esp_code_alternatif,
           annee) %>% 
    summarise(n_pres = sum(effectif > 0,
                           na.rm = T)) %>% 
  ungroup() %>% 
  complete(annee,
           esp_code_alternatif,
           fill = list(n_pres = 0))

opes_par_an <- df1 %>% 
  group_by(annee) %>% 
    summarise(n_ope = n_distinct(ope_id)) %>% 
  ungroup()
  
df3 <- df3 %>% 
  left_join(y = opes_par_an) %>% 
  mutate(pc_pres = 100 * n_pres / n_ope)
```


# Sauvegarde des dataframes

```{r, echo = FALSE}
save(df1, file = "../processed_data/df1.RData")
save(df2, file = "../processed_data/df2.RData")
save(df3, file = "../processed_data/df3.RData")
```

```{r, eval = FALSE}
save(df1, file = "processed_data/df1.RData")
save(df2, file = "processed_data/df2.RData")
save(df3, file = "processed_data/df3.RData")
```
