==> scripts/40_construction_graphiques.Rmd <==
---
title: "Démonstrateur ASPE : Eléments Peuplement 1 point"
author: "Pascal Irz"
date: "`r format(Sys.time(), 'Le %d %B %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
---
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Chargement des packages et des données

```{r}
library(aspe)
library(tidyverse)
```


```{r, echo = FALSE}
load(file = "../processed_data/df1.RData")
load(file = "../processed_data/df2.RData")
load(file = "../processed_data/captures_et_env.RData")
load(file = "../raw_data/toutes_tables_aspe_sauf_mei.RData")

file.exists("../raw_data/toutes_tables_aspe_sauf_mei.RData")
```

```{r, eval = FALSE}
load(file = "processed_data/df1.RData")
load(file = "processed_data/df2.RData")
load(file = "processed_data/captures_et_env.RData")
load(file = "raw_data/toutes_tables_aspe_sauf_mei.RData")
```


# Définition de la palette de couleurs

```{r}
coul_contrast <- RColorBrewer::brewer.pal(n = 8, name = "Dark2")
coul_pastel <- RColorBrewer::brewer.pal(n = 8, name = "Pastel1")

mon_theme <-
  theme(
    panel.background = element_rect(fill = coul_pastel[2]),
    panel.grid.major = element_line(colour = coul_pastel[3]),
    panel.grid.minor = element_line(colour = coul_pastel[3],
                                    linetype = 'dotted',
                                    size = 0.7)
  )
```

# Chronique sur un point

```{r}
mon_pop <- "10603"
capt_mef_colo_ext <- mef_colo_ext_pops(df = df1,
                                       id_point = mon_pop)
```



```{r}
gg_colo_ext_pops(capt_mef_colo_ext)
```

# Facteurs environnementaux

## En moyenne

```{r}
# esp_communes <- captures_et_env %>%
#   group_by(esp_code_alternatif) %>% 
#     summarise(n_pres = sum(presence)) %>% 
#   ungroup() %>% 
#   filter(n_pres > 1000) %>% 
#   pull(esp_code_alternatif) %>% 
#   unique %>% 
#   as.character
# 
# captures_et_env <- captures_et_env %>% 
#   filter(esp_code_alternatif %in% esp_communes)
```

Pour visualiser l'influence des facteurs environnementaux sur la distribution des espèces, on peut comparer leur valeur moyenne entre les sites où l'espèce est présente et ceux où elle est absente :

```{r, fig.width = 12, fig.height = 12}
plots <- captures_et_env %>%
  gg_density_env_esp(espece = "BRE")

ggpubr::ggarrange(plotlist = plots,
                  ncol = 3, nrow = 3,
                  common.legend = TRUE)
```




```{r, fig.width = 12, fig.height = 12}
gg_evol_env_esp <- function(df, parametre, coul_pres = "blue", coul_abs = "red")
  
{
  
df %>% 
  filter(parametre == parametre) %>% 
  mutate(presence = ifelse(presence, "Pr\u00e9sence", "Absence")) %>% 
  ggplot(aes(x = annee, y = valeur_parametre, col = presence)) +
    geom_smooth(method = "lm") +
    labs(x = "", y = parametre, col = "") +
    scale_x_continuous(limits = c(1995, NA)) +
    scale_y_continuous(limits = c(0, NA)) +
    scale_fill_manual(values = c(coul_abs, coul_pres)) +
    facet_wrap(~esp_code_alternatif,
               scales = "free_y", ncol = 5)
  
}

parametres <- captures_et_env %>% 
  pull(parametre) %>% 
  unique %>% 
  as.character

captures_et_env2 %>% 
  gg_evol_env_esp(parametre = parametres[1])

prov <- captures_et_env2 %>%
  select(esp_code_alternatif, annee, presence, parametre, valeur_parametre) %>% 
  distinct() %>% 
  group_by(esp_code_alternatif, annee, presence, parametre) %>% 
    summarise(moyenne = mean(valeur_parametre, na.rm = T)) %>% 
  pivot_wider(names_from = presence, values_from = moyenne)

# map(.x = parametres[1],
#     .f = gg_evol_env_esp,
#     df = captures_et_env)

```

Les caractéristiques des sites prospectés ont-elles changé au fil des années ?

```{r}
env <- df1 %>%
         select(ope_id, annee) %>%
         distinct() %>%
  left_join(y = operation %>%
              select(ope_id, pop_id = ope_pop_id)) %>% 
  mef_ajouter_ope_env() %>% 
  pivot_longer(
    cols = distance_mer:temp_janvier,
    names_to = "parametre",
    values_to = "valeur_parametre",
    values_drop_na = TRUE
  )

env %>% 
  ggplot(aes(x = annee, y = valeur_parametre)) +
    #geom_boxplot(coef=1e30) +
   geom_smooth(method = "gam") +
    labs(x = "") +
    scale_x_continuous(limits = c(1990, NA)) +
    facet_wrap(~parametre,
               scales = "free_y",
               ncol = 4)
```


```{r}

```
